+++ CHANGELOG.md
@@ -7,7 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [[2.1.1](https://github.com/ElrondNetwork/dapp-core/pull/448)] - 2022-10-20
+- [Fixed transaciton signing with extra provider](https://github.com/ElrondNetwork/dapp-core/pull/444)
+- [Changed TimeAgo `div` tag to `span`](https://github.com/ElrondNetwork/dapp-core/pull/445)
+- [Updated WalletConnect provider to 2.1.0-beta.2, added ping helper, show only latest pairing per app](https://github.com/ElrondNetwork/dapp-core/pull/446)
+
+
+
 ## [[2.1.0](https://github.com/ElrondNetwork/dapp-core/pull/443)] - 2022-10-14
+
 - [Fix Trim Component on Safari on iOS 16](https://github.com/ElrondNetwork/dapp-core/pull/435)
 - [Removed moment dependency](https://github.com/ElrondNetwork/dapp-core/pull/439)
 - [Fixed type definitions](https://github.com/ElrondNetwork/dapp-core/pull/406)